+++ SDK/Engine/Chips/Game/GameChip.cs
@@ -292,7 +292,7 @@ namespace PixelVision8.Engine.Chips
         ///     This method returns the current background color ID. If no color exists, it returns -1
         ///     which is magenta (#FF00FF).
         /// </returns>
-        public virtual int BackgroundColor(int? id = null)
+        public virtual byte BackgroundColor(byte? id = null)
         {
             if (id.HasValue) ColorChip.backgroundColor = id.Value;
