+++ SDK/Lua/Chips/Game/LuaGameChip.cs
@@ -62,7 +62,7 @@ namespace PixelVision8.Engine.Chips
 
             #region Color APIs
 
-            LuaScript.Globals["BackgroundColor"] = new Func<int?, int>(BackgroundColor);
+            LuaScript.Globals["BackgroundColor"] = new Func<byte?, byte>(BackgroundColor);
             LuaScript.Globals["Color"] = new Func<int, string, string>(Color);
             LuaScript.Globals["ColorsPerSprite"] = new Func<int>(ColorsPerSprite);
             LuaScript.Globals["TotalColors"] = new Func<bool, int>(TotalColors);
